requirements.txt
torch
transformers
gradio
faiss-cpu
Pillow
numpy
datasets
requests